a2a_friend_scheduling/host_agent_adk/pyproject.toml
# This file defines the workspace for the a2a-friend-scheduling project.
# It tells uv that this project is a monorepo containing multiple packages.

[project]
name = "a2a-friend-scheduling"
version = "0.1.0"
description = "A multi-agent system for scheduling pickleball games."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Shared ADK & A2A Dependencies
    "google-adk>=1.2.1",
    "a2a-sdk>=0.2.5",
    "nest-asyncio>=1.6.0",
    "python-dotenv",
    "click",
    "google-generativeai",
    "httpx",
    # MOQT transport dependencies
    "aioquic>=1.2.0",
    "cryptography>=42.0.0",
    # Kaitlyn's agent dependencies (future)
    # "langgraph"
]
